fix: update .env to include AUTH_EXPIRY for jwt token generation
